pyproject.toml
[project]
name = "mixseek-quant-insight-plus"
version = "0.1.0"
description = "ClaudeCode integration for mixseek-quant-insight"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.13.7"
dependencies = [
    "mixseek-plus",
    "mixseek-quant-insight",
]

[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.7",
]

[tool.uv]
package = true

[tool.uv.build-backend]
module-name = "quant_insight_plus"

[tool.uv.sources]
mixseek-plus = { path = "../mixseek-plus", editable = true }
mixseek-quant-insight = { path = "../mixseek-quant-insight", editable = true }

[tool.ruff]
line-length = 119
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
plugins = ["pydantic.mypy"]
explicit_package_bases = true
mypy_path = ["src"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[[tool.mypy.overrides]]
module = "duckdb"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "quant_insight.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
warn_no_return = false

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
